fix(responses): preserve stateful history after executor refactor
